# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Simple POS API
  version: 1.0.0
  description: |-
    Contract-first API for Simple POS (ADR-0008). Precedence: BUSINESS_RULES.md (BR v1.2) > this file > the UI/UX PDF > code.

    **Hosts**: API `https://api.cashier.hossam-ameen.online`, PWA `https://cashier.hossam-ameen.online` (separate origins,
    CORS limited to the PWA origin, ADR-0003).

    **Conventions** (ADR-0005, 0008, 0010, 0013)
    - Paths have no trailing slash. JSON is snake_case. Ids are integers. Timestamps are RFC 3339 UTC.
    - **Money is always an integer number of minor units (piastres) in a field ending in `_minor`** (BR-GEN-01). The server
      computes every total, change, expected cash and difference (BR-GEN-03).
    - Success body: `{success: true, message, data}`; lists: `data = {items, pagination, …}`; 204 has no body.
    - Error body: `{success: false, code, message, details}`; `code` is a BR §9 code with its BR §9 HTTP status and Arabic
      `message` (BR-GEN-06).
    - Auth: `Authorization: Bearer <access_token>` (15 min). The refresh token is an httpOnly cookie on the API host only,
      used by `/api/auth/refresh` and `/api/auth/logout` with `X-Requested-With: simple-pos` (ADR-0003, ADR-0011).
    - `x-roles` lists the roles allowed to call each operation (`PUBLIC`, `ANY_AUTHENTICATED`, `ADMIN`, `CASHIER`); every
      other authenticated role gets `FORBIDDEN_ROLE` (BR-ROLE-05). `x-br` lists the rules the operation enforces and
...

...
servers:
...
security:
- bearerAuth: []
...
paths:
...
  /api/settings:
    get:
      operationId: getSettings
      tags:
      - settings
      summary: Read Settings
      description: |-
        The Settings singleton (business_name, receipt_footer). Any authenticated user may read it because the receipt uses it (BR-SET-02, BR-PAY-07).

        BR: BR-SET-01, BR-SET-02, BR-SET-03.
        Errors: `UNAUTHENTICATED`.
      x-roles:
      - ANY_AUTHENTICATED
      x-br:
      - BR-SET-01
      - BR-SET-02
      - BR-SET-03
      x-error-codes:
      - UNAUTHENTICATED
      responses:
        '200':
          description: Settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
    patch:
      operationId: updateSettings
      tags:
      - settings
      summary: Update Settings (الإعدادات)
      description: |-
        ADMIN only (BR-SET-02). business_name trimmed and non-empty (BR-SET-01); receipt_footer may be empty. There is no create/delete, no currency and no logo (BR-SET-03).

        BR: BR-SET-01, BR-SET-02, BR-SET-03.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `VALIDATION_ERROR`.
      x-roles:
      - ADMIN
      x-br:
      - BR-SET-01
      - BR-SET-02
      - BR-SET-03
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - VALIDATION_ERROR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsUpdate'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '422':
          $ref: '#/components/responses/Error422'
...
  /api/shifts:
    get:
      operationId: listShifts
      tags:
      - shifts
      summary: Shift history (screen 17, admin الورديات)
      description: |-
        Newest `opened_at` first. ADMIN: all shifts, optional `cashier_id` filter (US-29). CASHIER: only their own shifts
        (BR-ROLE-04); a `cashier_id` other than the caller → `FORBIDDEN_ROLE`. CLOSED rows use the stored snapshot
        (BR-SHIFT-08); OPEN rows use live totals (BR-SHIFT-05). `date_from`/`date_to` filter on `opened_at` (Cairo dates).

        BR: BR-ROLE-01, BR-ROLE-04, BR-SHIFT-05, BR-SHIFT-08, OQ-12, OQ-35.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`.
      x-roles:
      - ADMIN
      - CASHIER
      x-br:
      - BR-ROLE-01
      - BR-ROLE-04
      - BR-SHIFT-05
      - BR-SHIFT-08
      - OQ-12
      - OQ-35
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      parameters:
      - name: cashier_id
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/ShiftStatus'
      - $ref: '#/components/parameters/DateFrom'
      - $ref: '#/components/parameters/DateTo'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: A page of shifts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftPageEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
    post:
      operationId: openShift
      tags:
      - shifts
      summary: Open my shift (screen 03)
      description: |-
        Opens a shift for the caller with `opening_balance_minor` ≥ 0; `opened_at` is server time and any client value is
        ignored (BR-SHIFT-02). A second OPEN shift → `SHIFT_ALREADY_OPEN`, guaranteed by the partial unique index even under
        concurrent requests (BR-SHIFT-01, AC-03, ADR-0006). ADMIN → `FORBIDDEN_ROLE` (BR-ROLE-02).

        BR: BR-SHIFT-01, BR-SHIFT-02, BR-ROLE-02, BR-ROLE-03, AC-03, AC-11.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `SHIFT_ALREADY_OPEN`, `VALIDATION_ERROR`.
      x-roles:
      - CASHIER
      x-br:
      - BR-SHIFT-01
      - BR-SHIFT-02
      - BR-ROLE-02
      - BR-ROLE-03
      - AC-03
      - AC-11
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - SHIFT_ALREADY_OPEN
      - VALIDATION_ERROR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftOpen'
      responses:
        '201':
          description: Opened.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
  /api/shifts/current:
    get:
      operationId: getCurrentShift
      tags:
      - shifts
      summary: My current OPEN shift (screens 02, 04, 05)
      description: |-
        The caller's OPEN shift with live totals, expected cash and open order numbers, or `data: null` when none (BR-SHIFT-03 routing to screen 03).

        BR: BR-SHIFT-03, BR-SHIFT-05, BR-SHIFT-06, BR-SHIFT-04, BR-GEN-03, AC-06.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`.
      x-roles:
      - CASHIER
      x-br:
      - BR-SHIFT-03
      - BR-SHIFT-05
      - BR-SHIFT-06
      - BR-SHIFT-04
      - BR-GEN-03
      - AC-06
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      responses:
        '200':
          description: The OPEN shift or null.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentShiftEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
  /api/shifts/{id}:
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      operationId: getShift
      tags:
      - shifts
      summary: Shift detail / closed-shift summary (screen 05 read-only layout)
      description: |-
        OPEN: live totals. CLOSED: the immutable snapshot (BR-SHIFT-08). ADMIN: any shift. CASHIER: own shifts only; another
        cashier's existing shift → `FORBIDDEN_ROLE` (US-04.6, BR-ROLE-04 — GATE B default), unknown id → `NOT_FOUND`.

        BR: BR-ROLE-04, BR-SHIFT-05, BR-SHIFT-06, BR-SHIFT-07, BR-SHIFT-08, BR-GEN-03, OQ-12.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`.
      x-roles:
      - ADMIN
      - CASHIER
      x-br:
      - BR-ROLE-04
      - BR-SHIFT-05
      - BR-SHIFT-06
      - BR-SHIFT-07
      - BR-SHIFT-08
      - BR-GEN-03
      - OQ-12
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      responses:
        '200':
          description: The shift.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
  /api/shifts/{id}/close:
    parameters:
    - $ref: '#/components/parameters/Id'
    post:
      operationId: closeShift
      tags:
      - shifts
      summary: Close my shift (screen 05)
      description: |-
        In one transaction with the shift row locked (ADR-0006): the shift must be the caller's own (another cashier's →
        `FORBIDDEN_ROLE`) and OPEN (a CLOSED shift is terminal → `NO_OPEN_SHIFT`, GATE B default); any OPEN order → `SHIFT_HAS_OPEN_ORDERS` with `details.open_order_numbers` (BR-SHIFT-04). The server
        recomputes sales/cash/card totals over PAID orders only (BR-SHIFT-05), expected cash = opening + cash (BR-SHIFT-06),
        difference = counted − expected (BR-SHIFT-07), stores the snapshot and sets CLOSED forever (BR-SHIFT-08). The PWA then
        shows the summary and logs out (BR-SHIFT-09). ADMIN → `FORBIDDEN_ROLE` (OQ-3).

        BR: BR-SHIFT-04, BR-SHIFT-05, BR-SHIFT-06, BR-SHIFT-07, BR-SHIFT-08, BR-SHIFT-09, BR-ROLE-03, BR-GEN-03, OQ-3, AC-02, AC-03, AC-06.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`, `NO_OPEN_SHIFT`, `SHIFT_HAS_OPEN_ORDERS`, `VALIDATION_ERROR`.
      x-roles:
      - CASHIER
      x-br:
      - BR-SHIFT-04
      - BR-SHIFT-05
      - BR-SHIFT-06
      - BR-SHIFT-07
      - BR-SHIFT-08
      - BR-SHIFT-09
      - BR-ROLE-03
      - BR-GEN-03
      - OQ-3
      - AC-02
      - AC-03
      - AC-06
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      - NO_OPEN_SHIFT
      - SHIFT_HAS_OPEN_ORDERS
      - VALIDATION_ERROR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftClose'
      responses:
        '200':
          description: Closed; the stored summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
...
  /api/orders/{id}/payment:
    parameters:
    - $ref: '#/components/parameters/Id'
    post:
      operationId: payOrder
      tags:
      - payments
      summary: Pay an order — one payment, CASH or CARD (screen 13)
      description: |-
        One DB transaction (BR-PAY-04, ADR-0006/0007): lock the order row → idempotency lookup (replay or
        `IDEMPOTENCY_CONFLICT`) → ownership (`NOT_ORDER_OWNER`, BR-ROLE-06) → caller has an OPEN shift and the order belongs to
        it (`NO_OPEN_SHIFT`, BR-SHIFT-03) → order OPEN (`ORDER_NOT_EDITABLE`, BR-ORD-08) → method rules (`INSUFFICIENT_CASH`)
        → insert the single Payment (unique per order, BR-PAY-01) → set PAID with `paid_at`. The table becomes AVAILABLE
        (BR-PAY-05, BR-TBL-05). Before the lock: `UNAUTHENTICATED` → `FORBIDDEN_ROLE` → `NOT_FOUND` → `VALIDATION_ERROR`
        (body shape, missing/malformed `Idempotency-Key`, negative amount).
        - CASH: `amount_received_minor` required and ≥ total else `INSUFFICIENT_CASH`; change = received − total (BR-PAY-02).
        - CARD: `amount_received_minor` must be absent/null (else `VALIDATION_ERROR`); received = total, change = 0 (BR-PAY-03).
        Idempotency (`Idempotency-Key`, required, per user): same key + same order + same body → the original 201 response
        replayed with header `Idempotent-Replayed: true` and no second payment; same key with a different body **or a different
        order** → `IDEMPOTENCY_CONFLICT`; a different key on a PAID order → `ORDER_NOT_EDITABLE` (AC-07). Rejected attempts are not stored (ADR-0007). ADMIN → `FORBIDDEN_ROLE`.

        BR: BR-PAY-01, BR-PAY-02, BR-PAY-03, BR-PAY-04, BR-PAY-05, BR-TBL-05, BR-ORD-08, BR-SHIFT-03, BR-ROLE-02, BR-ROLE-06, BR-GEN-03, AC-01, AC-07, AC-08, AC-11.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`, `NOT_ORDER_OWNER`, `NO_OPEN_SHIFT`, `ORDER_NOT_EDITABLE`, `INSUFFICIENT_CASH`, `IDEMPOTENCY_CONFLICT`, `VALIDATION_ERROR`.
      x-roles:
      - CASHIER
      x-br:
      - BR-PAY-01
      - BR-PAY-02
      - BR-PAY-03
      - BR-PAY-04
      - BR-PAY-05
      - BR-TBL-05
      - BR-ORD-08
      - BR-SHIFT-03
      - BR-ROLE-02
      - BR-ROLE-06
      - BR-GEN-03
      - AC-01
      - AC-07
      - AC-08
      - AC-11
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      - NOT_ORDER_OWNER
      - NO_OPEN_SHIFT
      - ORDER_NOT_EDITABLE
      - INSUFFICIENT_CASH
      - IDEMPOTENCY_CONFLICT
      - VALIDATION_ERROR
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreate'
      responses:
        '201':
          description: Paid (first time or replay).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResultEnvelope'
          headers:
            Idempotent-Replayed:
              schema:
                type: string
                enum:
                - 'true'
              description: Present only when the stored response is replayed.
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
  /api/orders/{id}/receipt:
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      operationId: getOrderReceipt
      tags:
      - payments
      summary: Receipt data for printing (screen 14)
      description: |-
        Receipt fields in BR-PAY-07 order, for the 80 mm browser print layout (BR-PAY-06). Only the owning CASHIER
        (`NOT_ORDER_OWNER` otherwise); ADMIN → `FORBIDDEN_ROLE`; an order that is not PAID → `NOT_FOUND` (no receipt exists).
        Ownership is checked before the PAID check.
        Read-only, so refreshing screen 14 never re-submits payment (US-14.5). Printing is offered only on screen 14 (OQ-10).

        BR: BR-PAY-06, BR-PAY-07, BR-SET-01, BR-ROLE-06, OQ-1, OQ-8, OQ-9, OQ-10.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`, `NOT_ORDER_OWNER`.
      x-roles:
      - CASHIER
      x-br:
      - BR-PAY-06
      - BR-PAY-07
      - BR-SET-01
      - BR-ROLE-06
      - OQ-1
      - OQ-8
      - OQ-9
      - OQ-10
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      - NOT_ORDER_OWNER
      responses:
        '200':
          description: Receipt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: pos_refresh
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    DateFrom:
      name: date_from
      in: query
      required: false
      description: Inclusive Cairo calendar date (ADR-0013).
      schema:
        type: string
        format: date
    DateTo:
      name: date_to
      in: query
      required: false
      description: Inclusive Cairo calendar date (ADR-0013).
      schema:
        type: string
        format: date
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: BR-PAY-04 / ADR-0007. 8–128 chars `[A-Za-z0-9_-]` (the PWA sends a UUID v4 per attempt); missing or malformed
        → VALIDATION_ERROR. Scoped per user; the request hash covers the path order id and the body, so reusing a key on another
        order or with another body → IDEMPOTENCY_CONFLICT.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: ^[A-Za-z0-9_-]{8,128}$
...
  responses:
    Error401:
      description: INVALID_CREDENTIALS (login) or UNAUTHENTICATED (missing/invalid/expired/revoked session).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error403:
      description: FORBIDDEN_ROLE, NOT_ORDER_OWNER, or USER_DISABLED (login).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error404:
      description: NOT_FOUND.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error409:
      description: State conflict — see the operation's x-error-codes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error422:
      description: VALIDATION_ERROR, ORDER_EMPTY, DISCOUNT_INVALID or INSUFFICIENT_CASH.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    ShiftStatus:
      type: string
      enum:
      - OPEN
      - CLOSED
...
    PaymentMethod:
      type: string
      enum:
      - CASH
      - CARD
      description: BR-PAY-01.
    ErrorCode:
      type: string
      description: BR §9 codes (BR-GEN-06). The HTTP status of each code is fixed by BR §9. `INTERNAL_ERROR` (500) is technical,
        not a business rejection; approved by the owner at GATE B (D3) and added to BR §9 in BR v1.3 (ADR-0010).
      enum:
      - INVALID_CREDENTIALS
      - USER_DISABLED
      - TOO_MANY_ATTEMPTS
      - FORBIDDEN_ROLE
      - SHIFT_ALREADY_OPEN
      - NO_OPEN_SHIFT
      - SHIFT_HAS_OPEN_ORDERS
      - TABLE_OCCUPIED
      - TABLE_INACTIVE
      - ITEM_INACTIVE
      - ORDER_EMPTY
      - ORDER_NOT_EDITABLE
      - NOT_ORDER_OWNER
      - DISCOUNT_INVALID
      - INSUFFICIENT_CASH
      - ITEM_IN_USE
      - CATEGORY_NOT_EMPTY
      - USER_HAS_OPEN_SHIFT
      - LAST_ADMIN
      - VALIDATION_ERROR
      - UNAUTHENTICATED
      - NOT_FOUND
      - DUPLICATE_VALUE
      - IDEMPOTENCY_CONFLICT
      - INTERNAL_ERROR
...
    Error:
      type: object
      required:
      - success
      - code
      - message
      - details
      properties:
        success:
          type: boolean
          enum:
          - false
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Arabic message from BR §9, verbatim.
        details:
          $ref: '#/components/schemas/ErrorDetails'
...
    Settings:
      type: object
      description: Singleton (BR-SET-01). Only these two fields exist (BR-SET-03, US-28.6).
      required:
      - business_name
      - receipt_footer
      properties:
        business_name:
          type: string
          minLength: 1
          maxLength: 80
        receipt_footer:
          type: string
          maxLength: 200
    SettingsUpdate:
      type: object
      properties:
        business_name:
          type: string
          minLength: 1
          maxLength: 80
          description: Trimmed; empty/whitespace → VALIDATION_ERROR (BR-SET-01). Max 80 = GATE B D2.
        receipt_footer:
          type: string
          maxLength: 200
          description: May be empty (BR-SET-01). Max 200 = GATE B D2.
...
    Shift:
      type: object
      required:
      - id
      - code
      - cashier_id
      - cashier_name
      - status
      - opened_at
      - closed_at
      - opening_balance_minor
      - sales_total_minor
      - cash_total_minor
      - card_total_minor
      - orders_count
      - open_orders_count
      - open_order_numbers
      - expected_cash_minor
      - counted_cash_minor
      - difference_minor
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        code:
          type: string
          description: '"SH-" + id zero-padded to 4 (OQ-35).'
        cashier_id:
          type: integer
          format: int64
          minimum: 1
        cashier_name:
          type: string
        status:
          $ref: '#/components/schemas/ShiftStatus'
        opened_at:
          type: string
          format: date-time
          description: RFC 3339 UTC (BR-GEN-02).
        closed_at:
          type: string
          format: date-time
          description: RFC 3339 UTC (BR-GEN-02).
          nullable: true
        opening_balance_minor:
          type: integer
          format: int64
          description: BR-SHIFT-02. Integer minor units (BR-GEN-01).
          minimum: 0
        sales_total_minor:
          type: integer
          format: int64
          description: Σ total of PAID orders (BR-SHIFT-05). Live if OPEN, snapshot if CLOSED. Integer minor units (BR-GEN-01).
        cash_total_minor:
          type: integer
          format: int64
          description: Σ total of PAID CASH orders (BR-SHIFT-05). Integer minor units (BR-GEN-01).
        card_total_minor:
          type: integer
          format: int64
          description: Σ total of PAID CARD orders (BR-SHIFT-05). Integer minor units (BR-GEN-01).
        orders_count:
          type: integer
          description: Count of PAID orders (BR-SHIFT-05).
        open_orders_count:
          type: integer
          description: OPEN orders in this shift (OQ-15).
        open_order_numbers:
          type: array
          items:
            type: integer
          description: OPEN order numbers (BR-SHIFT-04; screen 05 disables close).
        expected_cash_minor:
          type: integer
          format: int64
          description: opening_balance + cash_total (BR-SHIFT-06). Integer minor units (BR-GEN-01).
        counted_cash_minor:
          type: integer
          format: int64
          description: Entered on close (BR-SHIFT-07); null while OPEN. Integer minor units (BR-GEN-01).
          nullable: true
        difference_minor:
          type: integer
          format: int64
          description: counted − expected (BR-SHIFT-07); < 0 عجز, > 0 زيادة, 0 مطابق (OQ-36); null while OPEN. Integer minor
            units (BR-GEN-01).
          nullable: true
...
    ShiftPage:
      type: object
      required:
      - items
      - pagination
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ShiftSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ShiftOpen:
      type: object
      required:
      - opening_balance_minor
      properties:
        opening_balance_minor:
          type: integer
          format: int64
          description: ≥ 0 (BR-SHIFT-02). `opened_at` is always server time. Integer minor units (BR-GEN-01).
          minimum: 0
    ShiftClose:
      type: object
      required:
      - counted_cash_minor
      properties:
        counted_cash_minor:
          type: integer
          format: int64
          description: ≥ 0 (BR-SHIFT-07). The server recomputes every total (BR-GEN-03). Integer minor units (BR-GEN-01).
          minimum: 0
...
    PaymentCreate:
      type: object
      required:
      - method
      properties:
        method:
          $ref: '#/components/schemas/PaymentMethod'
        amount_received_minor:
          type: integer
          format: int64
          nullable: true
          minimum: 0
          description: 'CASH: required, ≥ 0 (negative → VALIDATION_ERROR, OQ-27), ≥ total else INSUFFICIENT_CASH (BR-PAY-02).
            CARD: must be null/absent, else VALIDATION_ERROR (BR-PAY-03). Integer minor units.'
    PaymentResult:
      type: object
      required:
      - payment
      - order
      properties:
        payment:
          $ref: '#/components/schemas/Payment'
        order:
          $ref: '#/components/schemas/Order'
...
    SettingsEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          $ref: '#/components/schemas/Settings'
...
    ShiftPageEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          $ref: '#/components/schemas/ShiftPage'
    ShiftEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          $ref: '#/components/schemas/Shift'
    CurrentShiftEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          allOf:
          - $ref: '#/components/schemas/Shift'
          nullable: true
...
    PaymentResultEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          $ref: '#/components/schemas/PaymentResult'
    ReceiptEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          $ref: '#/components/schemas/Receipt'
...